RDK-407, DELIA-32699 : Changed rate parameter from float to int
Reason for change: Changed rate parameter from float to
int to rectify float inequality.

Test Procedure: Refer RDK-407

Risks: Low

+++ gstaamp.h
@@ -97,7 +97,7 @@ struct _GstAamp
 	media_stream stream[2];
 	gboolean audio_enabled;
 	gchar *location;
-	float rate;
+	gint rate;
 	GMutex mutex;
 	GstAampStreamer* context;
 	GCond state_changed;
